fix(cli): merge overlays in `share` + honor explicit --api-url
Two issues from Codex review on f18b464:

P1: `vibe-replay share` was uploading raw replay.json without merging
the user's editor annotations from overlays.json. The editor server's
/api/publish/cloud route already does the merge dance, so the CLI was
silently producing stale cloud links for any replay that had been
edited.

Extracted `loadOverlays` and `sessionWithEffectiveContent` from server.ts
into a shared `overlays.ts` module, and added `publishCloudWithOverlays`
to publishers/cloud.ts that wraps publishCloud with the merge / write /
restore-original dance. Both the picker cloud handler and `share`
command now use it. The /api/publish/cloud server route also uses the
shared helper, dropping the duplicated logic.

P2: When `VIBE_REPLAY_API_URL` was preset in the shell, an explicit
`--api-url https://vibe-replay.com` was silently ignored because we
only mutated process.env when the value differed from DEFAULT_API_URL.
Dropped the option default (mention it in the description for help
text instead) and now mutate process.env whenever --api-url is passed.
